test: unify ARControllerNFT tests and enhance build file selection - only debug test fails
