+++ package.json
@@ -63,11 +63,11 @@
     "build-docker": "docker exec emscripten-jsartoolkitnft npm run build",
     "build-docker-no-libar": "docker exec emscripten-jsartoolkitnft npm run build-no-libar",
     "test": "npm run test:all",
-    "test:wasm": "cross-env BUILD_TARGET=artoolkitNFT_wasm.js karma start karma.conf.js",
-    "test:debug": "cross-env BUILD_TARGET=artoolkitNFT.debug.js karma start karma.conf.js",
-    "test:simd": "cross-env BUILD_TARGET=artoolkitNFT_wasm.simd.js karma start karma.conf.js",
     "test:min": "cross-env BUILD_TARGET=artoolkitNFT.min.js karma start karma.conf.js",
-    "test:all": "npm run test:wasm && npm run test:debug && npm run test:simd && npm run test:min",
+    "test:debug": "cross-env BUILD_TARGET=artoolkitNFT.debug.js karma start karma.conf.js",
+    "test:wasm": "cross-env BUILD_TARGET=artoolkitNFT_wasm.js karma start karma.conf.js",
+    "test:simd": "cross-env BUILD_TARGET=artoolkitNFT_wasm.simd.js karma start karma.conf.js",
+    "test:all": "npm run test:min && npm run test:debug && npm run test:wasm && npm run test:simd",
     "watch": "./node_modules/.bin/watch 'npm run build' ./js/",
     "format-check": "prettier --check .",
     "format": "prettier --write .",
